@jinchengchenghh 2024-10-31
Spark spill compares the prefix first, if it is equal, then compare the record. @FelixYBW 
https://github.com/apache/spark/blob/master/core/src/main/java/org/apache/spark/util/collection/unsafe/sort/UnsafeSorterSpillMerger.java#L33