@jinchengchenghh 2024-11-07
The red CI is not related to this PR, I open this issue https://github.com/facebookincubator/velox/issues/11467 to track it, could we merge this one? Thanks! @xiaoxmeng